Refactored using Flask Blueprint. Added script to recreate db.
