+++ recreate_users_db.sh
@@ -1,0 +1,1 @@
+docker-compose -f docker-compose-dev.yml run users python manage.py recreate-db
