+++ services/users/manage.py
@@ -1,13 +1,14 @@
 # services/users/manage.py
 
-from flask.cli import FlaskGroup
-
-from project import app
-
 import unittest
 
+from flask.cli import FlaskGroup
 
-cli = FlaskGroup(app)
+from project import create_app, db
+from project.api.models import User
+
+app = create_app()
+cli = FlaskGroup(create_app=create_app)
 
 # adds new command line command called recreate_db
 @cli.command()
