+++ services/users/project/tests/__init__.py
@@ -1,0 +1,1 @@
+# services/users/project/tests/__init__.py
