+++ services/users/project/tests/base.py
@@ -2,7 +2,9 @@
 
 from flask_testing import TestCase
 
-from project import app, db
+from project import create_app, db
+
+app = create_app()
 
 class BaseTestCase(TestCase):
     def create_app(self):
